[xsave] fix tests
