+++ .travis.yml
@@ -17,6 +17,8 @@ matrix:
     - env: DOCUMENTATION
       install: true
       script: ci/dox.sh
+    - env: RUSTFMT=On TARGET=x86_64-unknown-linux-gnu
+    - env: CLIPPY=On TARGET=x86_64-unknown-linux-gnu
   allow_failures:
     - env: RUSTFMT=On TARGET=x86_64-unknown-linux-gnu
     - env: CLIPPY=On TARGET=x86_64-unknown-linux-gnu
